Create script for easier debugging of the emulator instructions.
The scripts starts a qemu, gdb, and gingersnap process. It connects
the gdb process to qemu and steps through the target executable
one instruction at a time. For every instruction, all relevant
registers in gdb are compared to the registers in gingersnap. If
something differs, it halts and prints the program counters of both
gingersnap and the target process in gdb.